fixed comparison

+++ src/client/gui/pawsloading.cpp
@@ -163,7 +163,7 @@ void pawsLoadWindow::DrawAnim()
         }
     } 
     
-    for(int i = 0; i < positions.GetSize(); ++i)
+    for(size_t i = 0; i < positions.GetSize(); ++i)
     { 
         dot->Draw(positions[i].x, positions[i].y);
     }
